+++ Sources/Client/Model/User/UsersQuery.swift
@@ -20,7 +20,12 @@ public struct UsersQuery: Encodable {
     /// Filter conditions.
     public let filter: Filter
     /// Sort options, e.g. `.init("last_active", isAscending: false)`
-    public let sort: Sorting?
+    @available(*, deprecated, message: "Please use `sorting` property")
+    public var sort: Sorting? {
+        sorting.first
+    }
+    /// Sort options, e.g. `[.init("last_active", isAscending: false)]`
+    public let sorting: [Sorting]
     /// Used for paginating response.
     public let pagination: Pagination
     /// Query options, e.g. .presence
@@ -30,12 +35,12 @@ public struct UsersQuery: Encodable {
     ///
     /// - Parameters:
     ///   - filter: filter conditions, e.g. `"name".equal(to: "rover_curiosity")`
-    ///   - sort: sort options, e.g. `.init("last_active", isAscending: false)`
+    ///   - sorting: sort options array, order dependent., e.g. `[Sorting.init("last_active", isAscending: false)]`
     ///   - pagination: Pagination for query. Only supports `.limit` and `.offset`
     ///   - options: Query options, e.g. `.presence`
-    public init(filter: Filter, sort: Sorting? = nil, pagination: Pagination = [.usersPageSize], options: QueryOptions = []) {
+    public init(filter: Filter, sorting: [Sorting], pagination: Pagination = [.usersPageSize], options: QueryOptions = []) {
         self.filter = filter
-        self.sort = sort
+        self.sorting = sorting
         self.pagination = pagination
         self.options = options
         
@@ -57,11 +62,27 @@ public struct UsersQuery: Encodable {
         }
     }
     
+    /// Init a users query.
+    ///
+    /// - Parameters:
+    ///   - filter: filter conditions, e.g. `"name".equal(to: "rover_curiosity")`
+    ///   - sort: sort options, e.g. `.init("last_active", isAscending: false)`
+    ///   - pagination: Pagination for query. Only supports `.limit` and `.offset`
+    ///   - options: Query options, e.g. `.presence`
+    public init(filter: Filter, sort: Sorting? = nil, pagination: Pagination = [.usersPageSize], options: QueryOptions = []) {
+        if let sort = sort {
+            self.init(filter: filter, sorting: [sort], pagination: pagination, options: options)
+        } else {
+            self.init(filter: filter, sorting: [], pagination: pagination, options: options)
+        }
+    }
+    
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(filter, forKey: .filter)
         try container.encodeIfPresent(pagination.limit, forKey: .limit)
         try container.encodeIfPresent(pagination.offset, forKey: .offset)
+        try container.encode(sorting, forKey: .sort)
         try options.encode(to: encoder)
     }
 }