+++ Sources/Core/Client/Client+RxUser.swift
@@ -17,6 +17,19 @@ public extension Reactive where Base == Client {
     /// Requests users with given parameters. Creates a `UsersQuery` and call the `queryUsers` with it.
     /// - Parameters:
     ///   - filter: a user filter.
+    ///   - sorting: sorting options array.
+    ///   - pagination: Pagination for query. Only supports `.limit` and `.offset`
+    ///   - options: a query options.
+    func queryUsers(filter: Filter,
+                    sorting: [Sorting],
+                    pagination: Pagination = [.usersPageSize],
+                    options: QueryOptions = []) -> Observable<[User]> {
+        queryUsers(query: .init(filter: filter, sorting: sorting, pagination: pagination, options: options))
+    }
+    
+    /// Requests users with given parameters. Creates a `UsersQuery` and call the `queryUsers` with it.
+    /// - Parameters:
+    ///   - filter: a user filter.
     ///   - sort: a sorting.
     ///   - pagination: Pagination for query. Only supports `.limit` and `.offset`
     ///   - options: a query options.
@@ -24,7 +37,11 @@ public extension Reactive where Base == Client {
                     sort: Sorting? = nil,
                     pagination: Pagination = [.usersPageSize],
                     options: QueryOptions = []) -> Observable<[User]> {
-        queryUsers(query: .init(filter: filter, sort: sort, pagination: pagination, options: options))
+        if let sort = sort {
+            return queryUsers(query: .init(filter: filter, sorting: [sort], pagination: pagination, options: options))
+        } else {
+            return queryUsers(query: .init(filter: filter, sorting: [], pagination: pagination, options: options))
+        }
     }
     
     /// Requests users with a given query (see `UsersQuery`).